Move union_city pytest.fixture to conftest.py
