+++ tests/conftest.py
@@ -95,3 +95,23 @@ def inro_context(pytestconfig):
         except ModuleNotFoundError:
             print("Mocking inro environment.")
             mocked_inro_context()
+
+@pytest.fixture(scope="session")
+def union_city(examples_dir, root_dir, inro_context):
+    """Union City model run testing fixture."""
+    from tm2py.controller import RunController
+    from tm2py.examples import get_example
+
+    EXAMPLE = "UnionCity"
+    _example_root = os.path.join(examples_dir, EXAMPLE)
+
+    get_example(example_name="UnionCity", root_dir=root_dir)
+    controller = RunController(
+        [
+            os.path.join(examples_dir, "scenario_config.toml"),
+            os.path.join(examples_dir, "model_config.toml"),
+        ],
+        run_dir=_example_root,
+    )
+    controller.run()
+    return controller
